requirements.txt
pycryptodome
colorama
wmi